Added vulkan loaders as required.
This PR adds vulkan-icd-loader and lib32-vulkan-icd-loader which are required for Vulkan support 

https://wiki.archlinux.org/title/Vulkan#Installation

The reason for mentioning loaders is because if it is not mentioned it will install vulkan-icd-loader as a build dependency (make) in case of vulkan-radeon for example. And if someone removes the make dependencies vulkan-icd-loader gets removed as well.